update: changin de type of data on user migration, row phone

+++ src/database/migrations/1726684076550-users.ts
@@ -46,7 +46,7 @@ export class Users1726684076550 implements MigrationInterface {
                 },
                 {
                     name: 'phone',
-                    type: 'int',
+                    type: 'BIGINT',
                     isNullable: false
                 },
                 {
